Brought AROS grub2 in sync with grub2 repository revision 1919
git-svn-id: https://svn.aros.org/svn/aros/trunk@30086 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ arch/common/boot/grub2/commands/cmp.c
@@ -51,8 +51,8 @@ grub_cmd_cmp (struct grub_arg_list *state __attribute__ ((unused)),
 
   if (grub_file_size (file1) != grub_file_size (file2))
     grub_printf ("Differ in size: %llu [%s], %llu [%s]\n", 
-		 grub_file_size (file1), args[0], 
-		 grub_file_size (file2), args[1]);
+		 (unsigned long long) grub_file_size (file1), args[0],
+		 (unsigned long long) grub_file_size (file2), args[1]);
   else
     {
       pos = 0;
@@ -78,7 +78,7 @@ grub_cmd_cmp (struct grub_arg_list *state __attribute__ ((unused)),
 	      if (buf1[i] != buf2[i])
 		{
 		  grub_printf ("Differ at the offset %llu: 0x%x [%s], 0x%x [%s]\n",
-			       i + pos, buf1[i], args[0],
+			       (unsigned long long) (i + pos), buf1[i], args[0],
 			       buf2[i], args[1]);
 		  goto cleanup;
 		}
